qemu: configuration to use qemu debug console
QEMU provides a debug console for x86 CPUs which could be binded on
standard output. It really easy to use so it could be an easy way to
debug when display is not available yet. The support for this console is
now available in boot process as a callback from print library.

A flexible way to define kernel configuration is now available in build
chain: it is possible to define a configuration in kernel-configs.mk
which modify compilation flags and so on to fit new requirements when
the target is called.

New target are available in global makefile to directly one or another
emulator depending on which functionnality you need.