Architecture dependencies: introduce x86 real mode code.
Bioscall will be needed to probe memory on Intel x86 CPUs, so we'll need
a piece of code to switch the CPU to real mode and perform bios calls.
This changeset introduce a switch-to-real-mode code, a code to perform
bios call and all necessary stuff to separate architecture dependent and
independent implementation. The build toolchain is now able to manage
multiple implementations (architectures ?) for same interfaces.