Bioscall: adapt to GDT translation.
At early boot stage the kernel is in 3GB virtual memory and the
translation is done by special segments in the Global Descriptor Table.
As bioscall have to be done in real mode without virtual adress space,
the code have been adapted to enter into a proper context before, and
restore the virtual context after the BIOS call.