Stage2: retrieve information from previous stages
Provides a way to retrieve hardware/context information gathered by stages
before the stage 2. Indeed bootloader and bootstrap process provided and
installed a boot context before calling stage 2 and it is valuable to get
these information for further hardware initialization/manipulation.

The implementation is architecture dependant and is provided for x86 only.